autostart: Add retries
On system boot, the machined DBus service may not be started when
we attempt to start a container. This results in a 30 second timeout
on the machinectl command, and autostart would fail. Adding some retry
logic should ensure containers definitely do start on startup.

Also enabled on-failure restarts in Systemd for the autostart service.